XCrawler/requirements.txt
# X API V2 Tweet Monitor Dependencies

# HTTP requests for X API
requests>=2.31.0

# Environment variable management
python-dotenv>=1.0.0
